+++ Userland/SampleCodeModule/testSem.c
@@ -115,8 +115,8 @@ void test_sync(){
         printf("CREATING PROCESSES...(WITH SEM)\n");
 
         for(i = 0; i < TOTAL_PAIR_PROCESSES; i++){
-            newP(&incSemA);
-            newP(&incSemB);
+            newP((uint64_t)&incSemA);
+            newP((uint64_t)&incSemB);
         }
     }
     for (int j = 0; j < TOTAL_PAIR_PROCESSES; ++j) {
@@ -137,8 +137,8 @@ void test_no_sync(){
     if ((i = my_sem_open(AUX_SEM, 0)))
         printf("ERROR OPENING SEM: %d\n", i);
     for(i = 0; i < TOTAL_PAIR_PROCESSES; i++){
-        newP(&incNoSemA);
-        newP(&incNoSemB);
+        newP((uint64_t)&incNoSemA);
+        newP((uint64_t)&incNoSemB);
     }
     for (int j = 0; j < TOTAL_PAIR_PROCESSES; ++j) {
         my_sem_wait(AUX_SEM);
